test: move client/shared testHelpers to src/testing/
