+++ client/browser/src/shared/code-hosts/shared/views.test.ts
@@ -3,7 +3,7 @@ import { from, Observable, of, Subject, Subscription, NEVER } from 'rxjs'
 import { bufferCount, map, switchMap, toArray } from 'rxjs/operators'
 import * as sinon from 'sinon'
 
-import { createBarrier } from '@sourcegraph/shared/src/api/integration-test/testHelpers'
+import { createBarrier } from '@sourcegraph/shared/src/testing/testHelpers'
 
 import { MutationRecordLike } from '../../util/dom'
