+++ client/shared/src/api/integration-test/configuration.test.ts
@@ -1,9 +1,8 @@
 import { BehaviorSubject, of } from 'rxjs'
 
 import { EMPTY_SETTINGS_CASCADE, SettingsCascadeOrError } from '../../settings/settings'
+import { assertToJSON, integrationTestContext } from '../../testing/testHelpers'
 import { SettingsEdit } from '../client/services/settings'
-
-import { assertToJSON, integrationTestContext } from './testHelpers'
 
 describe('Configuration (integration)', () => {
     test('is synchronously available', async () => {
