+++ client/shared/src/api/integration-test/documents.test.ts
@@ -1,6 +1,5 @@
 import type { TextDocument } from '../../codeintel/legacy-extensions/api'
-
-import { assertToJSON, collectSubscribableValues, integrationTestContext } from './testHelpers'
+import { assertToJSON, collectSubscribableValues, integrationTestContext } from '../../testing/testHelpers'
 
 describe('Documents (integration)', () => {
     describe('workspace.textDocuments', () => {
