+++ client/shared/src/api/integration-test/search.test.ts
@@ -1,8 +1,7 @@
 import { take } from 'rxjs/operators'
 
+import { integrationTestContext } from '../../testing/testHelpers'
 import { wrapRemoteObservable } from '../client/api/common'
-
-import { integrationTestContext } from './testHelpers'
 
 describe('search (integration)', () => {
     test('registers a query transformer', async () => {
