+++ client/shared/src/api/integration-test/windows.test.ts
@@ -3,11 +3,10 @@ import { from, of } from 'rxjs'
 import { switchMap, take, toArray, first } from 'rxjs/operators'
 import type { ViewComponent, Window } from 'sourcegraph'
 
+import { assertToJSON, integrationTestContext } from '../../testing/testHelpers'
 import { wrapRemoteObservable } from '../client/api/common'
 import { NotificationType } from '../extension/extensionHostApi'
 import { TextDocumentData } from '../viewerTypes'
-
-import { assertToJSON, integrationTestContext } from './testHelpers'
 
 describe('Windows (integration)', () => {
     describe('app.activeWindow', () => {
