+++ client/shared/src/hover/actions.test.ts
@@ -17,8 +17,8 @@ import { ActionItemAction } from '../actions/ActionItem'
 import { ExposedToClient } from '../api/client/mainthread-api'
 import { FlatExtensionHostAPI } from '../api/contract'
 import { WorkspaceRootWithMetadata } from '../api/extension/extensionHostApi'
-import { integrationTestContext } from '../api/integration-test/testHelpers'
 import { PlatformContext, URLToFileContext } from '../platform/context'
+import { integrationTestContext } from '../testing/testHelpers'
 import {
     FileSpec,
     UIPositionSpec,
